+++ myapp/__init__.py
@@ -1,9 +1,9 @@
 __title__ = 'myapp'
-__version__ = '1.1.0'
+__version__ = '1.2.0'
 __author__ = 'Alan Crosswell'
 __license__ = 'Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0) License'
 __license_url__ = 'https://creativecommons.org/licenses/by-nc-sa/4.0/legalcode'
-__copyright__ = '(c) 2018-2020 The Trustees of Columbia University in the City of New York'
+__copyright__ = '(c) 2018-2021 The Trustees of Columbia University in the City of New York'
 
 # Version synonym
 VERSION = __version__